backend/reservations/views.py
from rest_framework import viewsets
from .models import Table, Reservation
from .serializers import TableSerializer, ReservationSerializer

# Create your views here.

class TableViewSet(viewsets.ModelViewSet):
    queryset = Table.objects.all()
    serializer_class = TableSerializer

class ReservationViewSet(viewsets.ModelViewSet):
    queryset = Reservation.objects.all()
    serializer_class = ReservationSerializer