backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
django-jazzmin==3.0.1
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.1
